Fix the `z-index` for Material UI Dialog
