+++ src/components/AppLayout/AppLayout.tsx
@@ -2,9 +2,10 @@ import useAppState from "@dashboard/hooks/useAppState";
 import { LinearProgress } from "@material-ui/core";
 import { useActionBar } from "@saleor/macaw-ui";
 import { Box } from "@saleor/macaw-ui/next";
-import React, { useContext, useState } from "react";
+import React, { useState } from "react";
 
-import { DevModeContext, DevModePanel } from "../DevModePanel/DevModePanel";
+import { DevModePanel } from "../DevModePanel/DevModePanel";
+import { useDevModeContext } from "../DevModePanel/hooks";
 import { useDevModeKeyTrigger } from "../DevModePanel/useDevModeKeyTrigger";
 import Navigator from "../Navigator";
 import { Sidebar } from "../Sidebar";
@@ -22,7 +23,7 @@ const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
   const [appState] = useAppState();
   const [isNavigatorVisible, setNavigatorVisibility] = useState(false);
 
-  const { isDevModeVisible, setDevModeVisibility } = useContext(DevModeContext);
+  const { isDevModeVisible, setDevModeVisibility } = useDevModeContext();
 
   useDevModeKeyTrigger(() => setDevModeVisibility(!isDevModeVisible));
